+++ CHANGES.md
@@ -7,6 +7,8 @@ Glass 0.6 — ????-??-??
   * The `$` prefix can now be used for hexadecimal numbers.
   * Symbols can no longer start with the `$` character (bc).
   * The `>>>` unsigned shift right operator is now supported.
+  * The `'` character can now be written as `''` in character literals.
+  * The `"` character can now be written as `""` in string literals.
 
 Glass 0.5 — 2017-01-18
 ----------------------